adding clang format to the project
Modified  template

+++ bin/dgraph.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
 # this script assumes that you are using Competitive Programming helper
-#
 if [ ! -d ~/projects/cp/.cph ]; then
   echo "./.cph does not exist. Use CPH for this script"
   exit
@@ -13,7 +12,8 @@ spinner() {
   sp="/-\|"
   echo -n ' '
   while [ -d /proc/$PID ]; do
-    printf "\b${sp:i++%${#sp}:1}"
+    sleep 0.2
+    printf "\r\b${sp:i++%${#sp}:1} Compiling"
   done
 }
 # compiling
@@ -25,7 +25,7 @@ if [ ! -f /tmp/graph.out ]; then
 else
   echo -en "\rCompilation Done\n"
 
-  mkdir -p ~/projects/cp/test.graph/
+  mkdir -p /tmp/test.graph/
 
   # read each item in the JSON array to an item in the Bash array
   array="$(cat ./.cph/"."$1* | jq -c -S '.tests[]')"
@@ -35,14 +35,23 @@ else
   for item in "${my_array[@]}"; do
     input="$(jq --raw-output '.input' <<<"$item")"
     id="$(jq --raw-output '.id' <<<"$item")"
-    echo "$input" >./test.graph/$1.$id.test
+    echo "$input" >/tmp/test.graph/$1.$id.test
   done
 
-  for file in ~/projects/cp/test.graph/$1.*; do
+  for file in /tmp/test.graph/$1.*; do
     [ -f "$file" ] && echo "Processing $file" && /tmp/graph.out <"$file"
   done
 
-  rm /tmp/graph.out
-  rm /tmp/graph.svg
-  rm /tmp/graph.dot
+  if [ -f /tmp/graph.out ]; then
+    rm /tmp/graph.out
+  fi
+
+  # if [ -f /tmp/graph.svg ]; then
+  #   rm /tmp/graph.svg
+  # fi
+
+  # if [ -f /tmp/graph.dot ]; then
+  #   rm /tmp/graph.dot
+  # fi
+
 fi
